Made a lot of changes.
Made start to a boss enemy in the game.
Added one new material for the boss.
Added a widget to the boss which displays health via progress bar.
Added health and score systems to Jon's added AI.
Added hit registration and a new projectile to be used by the turret AI's.
Likely missed some stuff but these were the main contributions.